use true-to-hardware sizing and remove Display struct
Removing the display struct to keep things simple. Less indirections.
Also made the data structure more true to the hardware. This does
increase the amount of time I need to cast when doing operations,
especially indexing. But that's just a consequence of emulating a
limited "hardware" system with Rust's default types.